vesa: Added VESA support
- Implemented driver for use VESA functionality
- Changed early_screen driver for using VESA
- Implemented v86 monitor for calling BIOS routines (needed by VESA)
- Added simply 8x8 font
- Removed unnecessary/garbage code
- Improved some structures
- Using MSR 0xC0000102 instead of 0xC0000101 for holding per-cpu structure (temporary?)
- Changed sleep_no_task function for using number of msec instead of ticks
- Scheduler fixes
- Move some functionality from ASM interrupt handler to high-level interrupt handler
- Fixed wrong parameter types in tss_entry_t, and added iomap bitmap

+++ include/sys/tss.h
@@ -29,8 +29,8 @@ typedef struct {
 	uint32_t	fs;
 	uint32_t	gs;
 	uint32_t	ldt;
-	uint32_t	trap;
-	uint32_t	iomap_base;
+	uint16_t	trap;
+	uint16_t	iomap_base;
 } __attribute__((packed)) tss_entry_t;
 
 void		tss_init();
